requirements.txt
alembic==1.4.2
amqp==2.6.1
attrs==19.3.0
billiard==3.6.4.0
blinker==1.4
Bcrypt==3.2.2
cached-property==1.5.1
cachetools==5.0.0
celery==4.4.6
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==8.0.1
colorama==0.4.3
configobj==5.0.6
cryptography==2.8
dbus-next>=0.0.0
distro==1.4.0
distro-info
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-JsonSchemaValidator==0.3.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
gevent==21.1.2
google-auth==2.6.0
gunicorn==20.1.0
greenlet==1.1.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
kombu==4.6.11
kubernetes==12.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.13.0
mongoengine==0.20.0
more-itertools==4.2.0
netaddr==0.8.0
netifaces==0.10.4
oauth2client==4.0.0
oauthlib==3.1.0
pika==1.2.1
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycryptodome==3.10.1
pymongo==3.12.3
PyMySQL==1.0.2
pyrsistent==0.15.5
python-dateutil==2.7.3
python-editor==1.0.4
pytz==2021.3
pyxdg==0.26
PyYAML>=3.10
rabbitmq==0.2.0
requests==2.24.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
s3transfer==0.3.7
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.18
urllib3==1.25.8
vine==1.3.0
websocket-client==0.53.0
Werkzeug==2.0.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
redis==4.3.1

boto3~=1.17.19
botocore~=1.20.112